# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Beauty Parlor API",
        "version": "v1"
    },
    "paths": {
        "/api/v1/user/{id}": {
            "put": {
                "tags": [
                    "Admin/Manage-users"
                ],
                "summary": "update user status",
                "description": "update user status",
                "operationId": "update user",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of user to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "update the status field",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Returns when user is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not authorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/users": {
            "get": {
                "tags": [
                    "Admin/Manage-users"
                ],
                "summary": "view all users",
                "description": "show all users",
                "operationId": "view all users",
                "responses": {
                    "401": {
                        "description": "Returns when user is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not authorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/create-shop": {
            "post": {
                "tags": [
                    "Shop"
                ],
                "summary": "create shop",
                "description": "shop user and invalidate token",
                "operationId": "create shop",
                "requestBody": {
                    "description": "fill the",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "shop_name",
                                    "shop_description",
                                    "shop_address",
                                    "opening_time",
                                    "closing_time"
                                ],
                                "properties": {
                                    "shop_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "test shop"
                                    },
                                    "shop_description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "test description"
                                    },
                                    "shop_address": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "test address"
                                    },
                                    "opening_time": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "8:00"
                                    },
                                    "closing_time": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "10:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Returns when user is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not authorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}